+++ firstSolar/front/src/ui-components/PooListCom320px.jsx
@@ -22,123 +22,9 @@ import {
 } from "../components/netdexlist/Netdexlist";
 import Poolitem320px from "../components/Pool/Poolitem320px";
 import { motion, LayoutGroup } from "framer-motion";
-import { dexList, getMainPoolList, netList } from "../api";
-import { useDispatch } from "react-redux";
-import { setIsLoading } from "../modules/isLoading.js";
-import {
-  mainNet768px1,
-  mainNet768px2,
-  platform768px1,
-  platform768px2,
-} from "../mainNet";
-import { useLocation } from "react-router-dom";
-
-const networkArray = [
-  "ethereum",
-  "optimism",
-  "metis",
-  "aurora",
-  "bsc",
-  "kava",
-  "heco",
-  "polygon",
-  "fantom",
-];
-
-const dexArray = [
-  "uniswap",
-  "pancakeswap",
-  "sushi",
-  "quickswap",
-  "linch",
-  "curve",
-  "bnt",
-  "knc",
-  "matcha",
-  "bal",
-];
 
 export default function PooListCom320px(props) {
   const { overrides, ...rest } = props;
-  const didMount = React.useRef(false);
-  const location = useLocation();
-  const queryParams = new URLSearchParams(location.search);
-
-  const [currentPagePoolList, setCurrentPagePoolList] = React.useState([]);
-  const [pageIndex, setPageIndex] = React.useState(
-    Number(queryParams.get("page")) || 1
-  );
-
-  const [filter, setFilter] = React.useState(String(queryParams.get("filter")));
-
-  const [totalPages, setTotalPages] = React.useState(1);
-  const [mainNetList, setMainNetList] = React.useState([]);
-  const [platformList, setPlatformList] = React.useState([]);
-  const [mainNetList1, setMainNetList1] = React.useState([]);
-  const [platformList1, setPlatformList1] = React.useState([]);
-
-  const dispatch = useDispatch();
-
-  React.useEffect(() => {
-    queryParams.set("page", pageIndex);
-    queryParams.set("filter", filter);
-
-    const newUrl = `${location.pathname}?${queryParams?.toString()}`;
-    window.history.replaceState(null, "", newUrl);
-  }, [pageIndex, location, queryParams]);
-
-  const getPoolList = async () => {
-    try {
-      dispatch(setIsLoading(true));
-      if (filter == "null") {
-        const { poolListData, resultTotalPages } = await getMainPoolList(
-          pageIndex
-        );
-        setCurrentPagePoolList(poolListData);
-        setTotalPages(resultTotalPages);
-      } else if (filter != "null") {
-        if (networkArray.includes(filter)) {
-          const data = await netList(filter, pageIndex);
-          setCurrentPagePoolList(data.poolListData);
-          setTotalPages(Math.ceil(data.poolListDataLength / 10));
-        } else if (dexArray.includes(filter)) {
-          const data = await dexList(filter, pageIndex);
-
-          setCurrentPagePoolList(data.poolListData);
-          setTotalPages(Math.ceil(data.poolListDataLength / 10));
-        }
-      }
-
-      setTimeout(() => {
-        dispatch(setIsLoading(false));
-      }, 5000);
-    } catch (error) {
-      dispatch(setIsLoading(false));
-      console.error(error);
-    }
-  };
-
-  React.useEffect(() => {
-    getPoolList();
-  }, [pageIndex]);
-
-  React.useEffect(() => {
-    if (didMount.current) setPageIndex(1);
-    else didMount.current = true;
-  }, [filter]);
-
-  React.useEffect(() => {
-    setMainNetList(Object.keys(mainNet768px1));
-    setPlatformList(Object.values(platform768px1));
-    setMainNetList1(Object.keys(mainNet768px2));
-    setPlatformList1(Object.values(platform768px2));
-  }, []);
-
-  const paginationProps = usePagination({
-    totalPages: totalPages,
-    currentPage: pageIndex,
-    siblingCount: 1,
-  });
 
   return (
     <View
@@ -245,15 +131,15 @@ export default function PooListCom320px(props) {
                 padding="0px 0px 0px 0px"
                 {...getOverrideProps(overrides, "Frame 3839412790")}
               >
-                {mainNetList?.map((item, idx) => (
+                {props?.mainNetList?.map((item, idx) => (
                   <Netlist320px
                     key={`Netlist320px-1${idx}`}
                     item={item}
-                    filter={filter}
-                    setFilter={setFilter}
-                    setCurrentPagePoolList={setCurrentPagePoolList}
-                    setTotalPages={setTotalPages}
-                    pageIndex={pageIndex}
+                    filter={props?.filter}
+                    setFilter={props?.setFilter}
+                    setCurrentPagePoolList={props?.setCurrentPagePoolList}
+                    setTotalPages={props?.setTotalPages}
+                    pageIndex={props?.pageIndex}
                   />
                 ))}
               </Flex>
@@ -272,15 +158,15 @@ export default function PooListCom320px(props) {
                 padding="0px 0px 0px 0px"
                 {...getOverrideProps(overrides, "Frame 3839412790")}
               >
-                {mainNetList1?.map((item, idx) => (
+                {props?.mainNetList1?.map((item, idx) => (
                   <Netlist320px
                     key={`Netlist320px-2${idx}`}
                     item={item}
-                    pageIndex={pageIndex}
-                    setCurrentPagePoolList={setCurrentPagePoolList}
-                    setTotalPages={setTotalPages}
-                    setFilter={setFilter}
-                    filter={filter}
+                    pageIndex={props?.pageIndex}
+                    setCurrentPagePoolList={props?.setCurrentPagePoolList}
+                    setTotalPages={props?.setTotalPages}
+                    setFilter={props?.setFilter}
+                    filter={props?.filter}
                   />
                 ))}
               </Flex>
@@ -363,14 +249,14 @@ export default function PooListCom320px(props) {
                 padding="0px 0px 0px 0px"
                 {...getOverrideProps(overrides, "Frame 3839412790")}
               >
-                {platformList?.map((item, idx) => (
+                {props?.platformList?.map((item, idx) => (
                   <Dexlist320px
                     key={`Dexlist320px-1${idx}`}
                     item={item}
-                    filter={filter}
-                    setCurrentPagePoolList={setCurrentPagePoolList}
-                    setTotalPages={setTotalPages}
-                    pageIndex={pageIndex}
+                    filter={props?.filter}
+                    setCurrentPagePoolList={props?.setCurrentPagePoolList}
+                    setTotalPages={props?.setTotalPages}
+                    pageIndex={props?.pageIndex}
                   />
                 ))}
               </Flex>
@@ -389,14 +275,14 @@ export default function PooListCom320px(props) {
                 padding="0px 0px 0px 0px"
                 {...getOverrideProps(overrides, "Frame 3839412790")}
               >
-                {platformList1?.map((item, idx) => (
+                {props?.platformList1?.map((item, idx) => (
                   <Dexlist320px
                     key={`Dexlist320px-2${idx}`}
                     item={item}
-                    filter={filter}
-                    setCurrentPagePoolList={setCurrentPagePoolList}
-                    setTotalPages={setTotalPages}
-                    pageIndex={pageIndex}
+                    filter={props?.filter}
+                    setCurrentPagePoolList={props?.setCurrentPagePoolList}
+                    setTotalPages={props?.setTotalPages}
+                    pageIndex={props?.pageIndex}
                   />
                 ))}
               </Flex>
@@ -410,7 +296,7 @@ export default function PooListCom320px(props) {
             initial={{ borderRadius: 25 }}
             transition={{ duration: 0.3, ease: [0.43, 0.13, 0.23, 0.96] }}
           >
-            {currentPagePoolList?.map((item, idx) => (
+            {props?.currentPagePoolList?.map((item, idx) => (
               <Poolitem320px
                 gap="17px"
                 direction="column"
@@ -431,15 +317,15 @@ export default function PooListCom320px(props) {
 
         <Flex width="88vw" justifyContent="center" padding="30px 0px 50px 0px">
           <Pagination
-            {...paginationProps}
+            {...props?.paginationProps}
             onChange={(pageNum) => {
               setPageIndex(pageNum);
             }}
             onNext={() => {
-              setPageIndex(pageIndex + 1);
+              setPageIndex(props?.pageIndex + 1);
             }}
             onPrevious={() => {
-              setPageIndex(pageIndex - 1);
+              setPageIndex(props?.pageIndex - 1);
             }}
           />
         </Flex>
